moves toaster to global/hooks
